Started to iterate over a "build reporter" class.
The BuildReporter responds to the raw build events by acting as the output translator's delegate.

It uses these events and internal logic to build up a report of the build, collect any metadata, errors etc. and keep track of the build state.

It in turn will eventually delegate to a formatter to display the build results as they progress.