Added translations for the clean command.
Something that cropped up as a result is that certain translations can both match lines in a
similar format (e.g. clean and build actions). To ensure this doesn't happen, the translations
should keep track of when it is valid for them to translate (e.g. the building translation
doesn't translate anything except the BUILD STARTED line until it knows a build has started).

The specs have been restructured slightly to reflect this stateful behaviour.